add support for the port field in SRV resources
This adds a new field to the Device struct that contains the `port` of
the service. This field defaults to `nil` but will be updated when it
receives the appropriate response.

to do this, a new `handle_device` is added to handle `:srv` type
resources. I opted to not include priority or weight fields as described
in the RFC since I'm not using them and have no way to test this, but
the variables have placeholders in case a future developer wants to
include this.